+++ src/pages/faq/FAQ.tsx
