+++ src/pages/faq/FAQDetail.tsx
