+++ src/pages/user/FindUserPassword.tsx
@@ -1,10 +1,10 @@
 import styled from "styled-components";
 import { Link } from "react-router-dom";
-import Footer from "../components/Footer";
-import TopNavBar from "../components/TopNavBar";
+import Footer from "../../components/Footer";
+import TopNavBar from "../../components/TopNavBar";
 import { Button, Form, FormGroup, Label } from "./Login";
 import { useState } from "react";
-import Input from "../components/Input";
+import Input from "../../components/Input";
 
 const Tab = styled.div`
   display: flex;
