+++ src/pages/user/Profile.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Input from "../components/Input";
+import Input from "../../components/Input";
 
 const Container = styled.div`
   display: flex;
